Codegen+Parser+Typechecker: Add early stages of codegen
There's also fixes on typechecker and on the parser I came across
when trying to add this feature.

New features added:
variable declaration
if/else statement
while statement
loop statement
functions (including external)
expression

The bugs I came accross are
the issues with scops as of result accidentelly removing variables
in current scope.
Also there's one where it doesn't save the return type on typechecker
now it's possible to modify Statements on typechecker to make it
easier to generate.